Moving testgroup example API to examples/
