+++ examples/apis/testgroup/register.go
