+++ examples/apis/testgroup/types.generated.go
