+++ examples/apis/testgroup/v1/register.go
