+++ examples/apis/testgroup/v1/types.generated.go
