+++ examples/apis/testgroup/v1/types.go
